spark/requirements.txt
py4j
pandas
psycopg2-binary
pymongo[srv]